Add 'KEEP_VCPKG_CACHE' to OBT-AWS01 settings and sequencer
Improve build.sh to check and if it  exists use then  the 'KEEP_VCPKG_CACHE'
to manage VCPKG cache.

+++ build.sh
@@ -267,8 +267,19 @@ fi
 
 if [ -d ~/.cache/vcpkg/archives ]
 then
-    WritePlainLog "Remove VCPKG leftovers." "${OBT_BUILD_LOG_FILE}"
-    rm -rf ~/.cache/vcpkg/archives
+    if [[ -v KEEP_VCPKG_CACHE ]]
+    then
+        if [[ ${KEEP_VCPKG_CACHE} -eq 1 ]]
+        then
+            WritePlainLog "Keep VCPKG leftovers." "${OBT_BUILD_LOG_FILE}"
+        else
+            WritePlainLog "Remove VCPKG leftovers." "${OBT_BUILD_LOG_FILE}"
+            rm -rf ~/.cache/vcpkg/archives
+        fi
+    else
+        WritePlainLog "Remove VCPKG leftovers." "${OBT_BUILD_LOG_FILE}"
+        rm -rf ~/.cache/vcpkg/archives
+    fi
 else
     WritePlainLog "There are not VCPKG leftovers." "${OBT_BUILD_LOG_FILE}"
 fi
